refactor: remove unused tests, components, and scripts from base package
- Deleted TodoItem component and its tests
- Removed TodoInput component and its tests
- Cleaned up vitest configuration and related dependencies
- Updated README to reflect removal of base tests
- Simplified todo logic by removing toggle functionality and related computed properties
