+++ packages/base/app/features/todo/composables/useTodos.ts
@@ -1,5 +1,5 @@
-import { ref, computed, readonly } from 'vue'
-import type { Todo } from '../TodoItem.vue'
+import { ref, readonly } from 'vue'
+import type { Todo } from '../TodoList.vue'
 
 export interface UseTodosOptions {
   supportCategories?: boolean
@@ -20,23 +20,13 @@ export function useTodos(options: UseTodosOptions = {}) {
     const newTodo: Todo = {
       id: generateId(),
       text: text.trim(),
-      completed: false,
     }
 
     // Only add category if supported and provided
     if (options.supportCategories && category?.trim()) {
       newTodo.category = category.trim()
     }
-
     todos.value.push(newTodo)
-  }
-
-  // Toggle todo completion status
-  const toggleTodo = (id: string) => {
-    const todo = todos.value.find(t => t.id === id)
-    if (todo) {
-      todo.completed = !todo.completed
-    }
   }
 
   // Delete a todo
@@ -47,48 +37,13 @@ export function useTodos(options: UseTodosOptions = {}) {
     }
   }
 
-  const completedTodos = computed(() => 
-    todos.value.filter(todo => todo.completed)
-  )
-
-  const pendingTodos = computed(() => 
-    todos.value.filter(todo => !todo.completed)
-  )
-
-  const totalCount = computed(() => todos.value.length)
-  const completedCount = computed(() => completedTodos.value.length)
-  const pendingCount = computed(() => pendingTodos.value.length)
-
-  // Get unique categories (only if categories are supported)
-  const categories = computed(() => {
-    if (!options.supportCategories) return []
-    
-    const categorySet = new Set<string>()
-    todos.value.forEach(todo => {
-      if (todo.category) {
-        categorySet.add(todo.category)
-      }
-    })
-    return Array.from(categorySet).sort()
-  })
-
-
   return {
     // State
     todos: readonly(todos),
     
     // Actions
     addTodo,
-    toggleTodo,
     deleteTodo,
-    
-    // Computed
-    completedTodos,
-    pendingTodos,
-    totalCount,
-    completedCount,
-    pendingCount,
-    categories,
    
   }
 }
