--- packages/ui/src/components/badge/Badge.vue
@@ -1,9 +1,0 @@
-<template>
-  <Badge v-bind="$attrs">
-    <slot />
-  </Badge>
-</template>
-
-<script setup lang="ts">
-import { Badge } from '@/components/ui/badge'
-</script>
