--- packages/ui/src/components/checkbox/Checkbox.vue
@@ -1,9 +1,0 @@
-<template>
-  <Checkbox v-bind="$attrs">
-    <slot />
-  </Checkbox>
-</template>
-
-<script setup lang="ts">
-import { Checkbox } from '@/components/ui/checkbox'
-</script>
