--- packages/ui/src/components/checkbox/index.ts
@@ -1,1 +1,0 @@
-export { default as Checkbox } from './Checkbox.vue'
