+++ packages/ui/src/index.ts
@@ -2,6 +2,4 @@ export * from './lib/cva-config-manager'
 
 // Custom Components
 export * from '@/components/button'
-export * from '@/components/badge'
-export * from '@/components/checkbox'
 export * from '@/components/input'
